+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased]
 ### Fixed
 - Use of undefined labels gives proper error message instead of Python exception.
+- `continue` in debugger doesn't get stuck on breakpoints.
 
 
 ## [0.4.0] - 2019-01-02